NFC: Moved GraphOperationInfo related code into a separate translation unit, in preparation for its new use in IRgen for decoding graph_op insts.
Elevated it from SILOptimizer/ to SIL/, so that IRGen can depend on it.
