+++ lib/SILOptimizer/Mandatory/TFUtilities.h
@@ -20,6 +20,7 @@
 
 #include "TFDeviceSupport.h"
 #include "swift/AST/TensorFlow.h"
+#include "swift/SIL/GraphOperationInfo.h"
 #include "swift/SIL/SILBuilder.h"
 #include "swift/SIL/SILFunction.h"
 #ifdef SWIFT_ENABLE_TENSORFLOW
@@ -107,74 +108,27 @@ struct SILTensorOpInfo {
   /// This is the TensorFlow name for the op.
   StringRef opName;
 
-  /// One of these records exists for every operand that the BuiltinInst has,
-  /// classifying the operand into a couple of buckets.  The most coarse grain
-  /// classification is "input" vs "attribute": the inputs come first,
-  /// followed by the attributes.  However, we need to be able to model the
-  /// fact that some input arguments are aggregated together into a single
-  /// input that is an array of tensors.  An integer attribute may be either
-  /// a Tensor value or an integer-encoded DType, etc.
-  enum class OperandClass {
-    /// Indicates one of the following:
-    /// 1) A normal tensor input: the value is a TensorHandle.
-    /// 2) An normal attribute (without modifier).
-    /// 3) A tensor or shape attribute (need a modifier for proper lowering).
-    /// 4) An array attribute (needed for parsing tfop, and dropped before graph
-    ///    lowering).
-    Input,
-
-    /// No modifier.
-    Normal,
-
-    /// Indicates that the array or scalar should be turned into a TF_Tensor.
-    Tensor,
-
-    /// Indicates that the array of integers should be interpreted as a shape.
-    Shape,
-
-    /// Indicates the metatype of a TensorFlow value type or an aggregate of
-    /// TensorFlow value types should be turned into a list of unknown shapes.
-    UnknownShapeList,
-
-    /// Indicates that the operand should be interpreted as an array. When
-    /// applied to the metatype of a TensorFlow value type or an aggregate of
-    /// TensorFlow value types, it will be flattened into an array of dtypes of
-    /// each TensorFlow value type as a Normal operand.
-    Array,
-
-    /// An operand specifying the address where an indirect output should be
-    /// stored.  This occurs when the tfop exists in a context where its output
-    /// is address-only.  Deabstraction eliminates Out operands before forming
-    /// graph_ops, by rewriting the tfop to return the value directly.  This
-    /// rewriting is possible because tfop outputs must always be loadable in
-    /// deabstraction scopes.
-    Out,
-  };
-
-  /// Return the string suffix for the specified attribute modifier.
-  static const char *getOperandClassSuffix(OperandClass opClass);
-
-  /// Return the operand class of the specified string form like "tensor"
-  static llvm::Optional<OperandClass> getOperandClass(StringRef suffix);
-
   /// These are the names of any attribute operands at the end of the list.
-  SmallVector<std::pair<StringRef, OperandClass>, 4> operandClasses;
+  SmallVector<std::pair<StringRef, GraphOperationInfo::OperandClass>, 4>
+      operandClasses;
 
   /// Return true if the specified operand is an input (not an attribute).
   bool isInput(unsigned operandNumber) const {
-    return operandClasses[operandNumber].second == OperandClass::Input;
+    return operandClasses[operandNumber].second ==
+           GraphOperationInfo::OperandClass::Input;
   }
 
   /// Returns the full name that this builtin would have if its operands
   /// changed to the passed-in values.
   std::string getBuiltinNameWithNewOperands(
-      const SmallVectorImpl<std::pair<StringRef, OperandClass>>
-        &newOperandClasses) const {
+      const SmallVectorImpl<
+          std::pair<StringRef, GraphOperationInfo::OperandClass>>
+          &newOperandClasses) const {
     std::string name = "__tfop_" + opName.str();
     for (const auto &newOperandClass : newOperandClasses) {
       name += ",";
       name += newOperandClass.first;
-      name += getOperandClassSuffix(newOperandClass.second);
+      name += GraphOperationInfo::getOperandClassSuffix(newOperandClass.second);
     }
     return name;
   }
@@ -187,68 +141,6 @@ struct SILTensorOpInfo {
 private:
   SILTensorOpInfo(BuiltinInst *inst) : inst(inst) {}
   bool decodeBuiltin();
-};
-
-/// Holds information about a TensorFlow operation as represented in SIL
-/// as GraphOperationInst.
-struct GraphOperationInfo {
-  /// The instruction being analyzed.
-  GraphOperationInst *inst;
-
-  explicit GraphOperationInfo(GraphOperationInst *inst) : inst(inst) {}
-
-  /// Return the device attribute associated with `inst`, which is required to
-  /// exist.
-  StringRef getDeviceString() const;
-
-  /// Return the device type for this instruction.
-  DeviceType getDeviceType() const {
-    return getOpDeviceType(getDeviceString());
-  }
-
-  enum InputMarker {
-    /// Scalar input, used by tfc.scalarToTensor only.
-    IM_Scalar,
-    /// Normal tensor, variant or resource input.
-    IM_Normal,
-    /// Marker for the start of an input list, has no corresponding operand.
-    IM_InputList,
-    /// Element of an input list.
-    IM_InputListElt,
-  };
-
-  /// Return a comma and letter identifier whose letter corresponds to the
-  /// specified InputMarker.
-  static const char *getInputMarker(InputMarker kind) {
-    switch (kind) {
-    case IM_Scalar:
-      return ",s";
-    case IM_Normal:
-      return ",i";
-    case IM_InputList:
-      return ",L";
-    case IM_InputListElt:
-      return ",e";
-    }
-  }
-
-  /// Decode the name of a graph_op into its TensorFlow op name and a list of
-  /// information about the operands.
-  StringRef decodeName(SmallVectorImpl<InputMarker> &inputInfo);
-
-  /// Given an attribute name like foo$tensor, decode the name and the class.
-  /// If there is no modifier specified, this defaults to
-  /// OperandClass::Normal.
-  static std::pair<StringRef, SILTensorOpInfo::OperandClass>
-  decodeAttributeName(Identifier name);
-
-  /// Get an int-typed attribute at `attrIdx`, which must have `attrName`.
-  int64_t getIntAttr(unsigned attrIdx, StringRef attrName) const;
-
-  /// Get a string-typed attribute at `attrIdx`, which must have `attrName`.
-  std::string getStringAttr(unsigned attrIdx, StringRef attrName) const;
-
-  void assertWithDump(bool cond, const char *assertMsg) const;
 };
 
 /// `inst` must have a single result, and return that result value.
